// ClientApp/src/app/models/employee.model.ts
export class Employee {
  employeeId?: number;
  firstName: string;
  maidenName?: string;
  lastName: string;
  contactNumber: string;
  alternativeNumber?: string;
  physicalAddress: string;
  postalAddress: string;
  said: string;
  email: string;
  username: string;
  password: string;
  imagePath: string;
  accessCode: string;
}