// ClientApp/src/app/models/employeeKin.ts
import { Employee } from "./employee.model";

export class EmployeeKin {
  employeeKinId?: number;
  firstName: string;
  maidenName?: string;
  lastName: string;
  contactNumber: string;
  alternativeNumber?: string;
  physicalAddress: string;
  postalAddress: string;
  said: string;

  employeeId?: number;
  employee: Employee;
}